自定义属性编辑OK

+++ Scripts/MapEditor/Property/MapSerializeData.cs
@@ -8,7 +8,7 @@ namespace MapSerializeData
     public class Prefab40SerializeData : Serializable.SerializeBase
     {
         public float edgebox_left;
-        public int edgebox_right;
+        public float edgebox_right;
         public float edgebox_down;
         public float edgebox_up;
     }
